refactor(sqlite): sqliteClient.ts shim を物理削除し offscreenGateway に一本化
- sqliteClient.ts (95l shim) を削除 — SqliteClient/getSharedSqliteClient は offscreenGateway に集約
- compositionManifest/auditLog/pipeline/types/storageMaintenance の import を offscreenGateway に移行
- テストの vi.mock('sqliteClient.js') を vi.mock('sqlite/offscreenGateway.js') に更新 (6 files)
- grep 'from.*sqliteClient' が 0件に

検証: type-check/lint/test 637 passed 11172 passed (サブセット検証済み)
