+++ src/background/__tests__/backgroundComposition.test.ts
@@ -35,7 +35,7 @@ const mocks = vi.hoisted(() => ({
 }));
 
 vi.mock('../obsidianClient.js', () => ({ ObsidianClient: mocks.ObsidianClient }));
-vi.mock('../sqliteClient.js', () => ({
+vi.mock('../sqlite/offscreenGateway.js', () => ({
   SqliteClient: mocks.SqliteClient,
   getSharedSqliteClient: mocks.getSharedSqliteClient,
 }));
