+++ src/background/__tests__/gistSyncTarget-r2.test.ts
@@ -8,7 +8,7 @@ const mockSet = vi.hoisted(() => vi.fn());
 
 import { GistSyncTarget } from '../syncTargets/gistSyncTarget.js';
 
-vi.mock('../sqliteClient.js', () => ({
+vi.mock('../sqlite/offscreenGateway.js', () => ({
   SqliteClient: vi.fn().mockImplementation(() => {
     const qr = vi.fn();
     const ur = vi.fn();
