+++ src/background/compositionManifest.ts
@@ -19,8 +19,8 @@ import { createAIService } from './ai/aiServiceFactory.js';
 import { RemoteAIService } from './ai/RemoteAIService.js';
 import type { AIService } from './ai/AIService.js';
 import { ObsidianClient } from './obsidianClient.js';
-import { getSharedSqliteClient } from './sqliteClient.js';
-import type { SqliteClient } from './sqliteClient.js';
+import { getSharedSqliteClient } from './sqlite/offscreenGateway.js';
+import type { SqliteClient } from './sqlite/offscreenGateway.js';
 import { setSqliteHealthCheck } from '../utils/storage/storageMaintenance.js';
 import { RecordingCacheInstance, SessionStoreRecordingCacheStore } from './recordingCache.js';
 import { TabCache } from './tabCache.js';
