+++ src/utils/__tests__/auditLog.test.ts
@@ -3,7 +3,7 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 const mockMutate = vi.fn().mockResolvedValue({ success: true, data: { id: 1 } });
 const mockQuery = vi.fn().mockResolvedValue({ success: true, data: { rows: [], total: 0 } });
 
-vi.mock('../../background/sqliteClient.js', () => {
+vi.mock('../../background/sqlite/offscreenGateway.js', () => {
   class MockSqliteClient {
     async mutate(op: Record<string, unknown>) {
       return mockMutate(op);
